+++ imio/dms/mail/migrations/migrate_to_3_0.py
@@ -689,22 +689,27 @@ class Migrate_To_3_0(Migrator):  # noqa
             rkey = "imio.pm.wsclient.browser.settings.IWS4PMClientSettings.field_mappings"
             rvalue = api.portal.get_registry_record(rkey, default=None)
             fns = [dic["field_name"] for dic in rvalue or []]
-            if fns and u"ignore_validation_for" not in fns:
-                fns.append(u"ignore_validation_for")
+            if fns:
+                if u"ignore_validation_for" not in fns:
+                    fns.append(u"ignore_validation_for")
+                    rvalue.append({"field_name": u"ignore_validation_for", "expression": u"string:groupsInCharge"})
+                if u"annexes" in fns:
+                    fns.remove(u"annexes")
+                    rvalue = [item for item in rvalue if item["field_name"] != u"annexes"]
                 orig_call = pm_item_data_vocabulary.__call__
-                pm_item_data_vocabulary.__call__ = lambda self0, ctxt: SimpleVocabulary([SimpleTerm(fn) for fn in fns])
-                rvalue.append({"field_name": u"ignore_validation_for", "expression": u"string:groupsInCharge"})
+                pm_item_data_vocabulary.__call__ = lambda self0, ctxt: SimpleVocabulary(
+                    [SimpleTerm(fn) for fn in fns]
+                )
                 api.portal.set_registry_record(rkey, rvalue)
                 pm_item_data_vocabulary.__call__ = orig_call
+
             # Rename IncomingmailWSClient to IncomingmailRestWSClient in imio.pm.wsclient field mappings
             orig_call = pm_item_data_vocabulary.__call__
-            pm_item_data_vocabulary.__call__ = lambda self0, ctxt: SimpleVocabulary([SimpleTerm(fn) for fn in fns])
             for field_mapping in rvalue:
                 field_mapping["expression"] = field_mapping["expression"].replace(
                     "@@IncomingmailWSClient", "@@IncomingmailRestWSClient"
                 )
             api.portal.set_registry_record(rkey, rvalue)
-            pm_item_data_vocabulary.__call__ = orig_call
 
             # imio.pm.wsclient
             self.portal.manage_permission(
@@ -723,7 +728,7 @@ class Migrate_To_3_0(Migrator):  # noqa
 
             # Uninsall imio.dms.soap2pm
             installer = api.portal.get_tool("portal_quickinstaller")
-            installer.uninstallProducts(["collective.js.tooltipster"])
+            installer.uninstallProducts(["imio.dms.soap2pm"])
 
             # END